Send: use coroutines and let room sending queues be independent of each others
